layout item

+++ app/src/main/java/com/marvel/heroes/domain/data/dto/Comics.java
@@ -12,7 +12,7 @@ public class Comics {
 
     public String title;
 
-    public long issueNumber;
+    public double issueNumber;
 
     public String variantDescription;
